split dynamodb lambda helper into sep file & dump rek_resp and logs as strings into dyndn instead of map
